python/interactive-chatbot/requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
memorystack>=1.0.0
python-dotenv>=1.0.0
